+++ tests/matrix/inverting_matrices/spotting_submatrices_test.c
@@ -5,18 +5,18 @@ MU_TEST(test_matrix_1_5_0_minus_3_2_7_0_6_minus_3_spotting_submatrix_is_minus_3_
 {
 	int		row = 0;
 	int		col = 2;
-	float	matrix_values[3][3] = {
+	double	matrix_values[3][3] = {
 		{1, 5, 0},
 		{-3, 2, 7},
 		{0, 6, -3}
 	};
-	float	expected_values[2][2] = {
+	double	expected_values[2][2] = {
 		{-3, 2},
 		{0, 6}
 	};
 
-	float **original_matrix = matrix(3);
-	float **expected = matrix(2);
+	double **original_matrix = matrix(3);
+	double **expected = matrix(2);
 	for(int i = 0; i < 3; i++)
 		for(int j = 0; j < 3; j++)
 			original_matrix[i][j] = matrix_values[i][j];
@@ -24,7 +24,7 @@ MU_TEST(test_matrix_1_5_0_minus_3_2_7_0_6_minus_3_spotting_submatrix_is_minus_3_
 		for (int j = 0; j < 2; j++)
 			expected[i][j] = expected_values[i][j];
 
-	float **result =  submatrix(original_matrix, row, col);
+	double **result =  submatrix(original_matrix, row, col);
 	mu_check(compare_matrices(result, expected));
 	free_matrix(original_matrix);
 	free_matrix(expected);
