+++ tests/tuples/operations/dot_product_test.c
@@ -5,8 +5,8 @@ MU_TEST(test_dot_product_of_two_tuples)
 {
 	t_tuple	v = vector(1, 2, 3);
 	t_tuple	w = vector(2, 3, 4);
-	float	expected = 20;
-	float	result = dot(v, w);
+	double	expected = 20;
+	double	result = dot(v, w);
 
 	mu_assert_double_eq(result, expected);
